Refactor : useTargetUesrPlaylists, useTargetUserProfileInfo에서  query와 service 로직 분리
